Separated power control from user space.
